upload final versions

+++ module4/ex02/calculator.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python3
 
-n1 = float(input("Enter the first number: "))
-n2 = float(input("Enter the second number: "))
+n1 = int(input("Enter the first number: "))
+n2 = int(input("Enter the second number: "))
 print("Thank you!")
 
 print(n1, '+', n2, '=', n1 + n2)
